Restructure project into scientific format: src layout, pyproject.toml, CITATION.cff, environment.yml, updated tests
