+++ src/echofoam_falsifiability/__init__.py
@@ -1,0 +1,6 @@
+"""
+Echofoam Falsifiability Suite package.
+
+This package contains simulation modules and utilities for testing
+Echofoam Theory under various scenarios.
+"""
