+++ tests/pyproject.toml
@@ -1,0 +1,18 @@
+[project]
+name = "echofoam-falsifiability-suite"
+version = "0.1.0"
+description = "A scientific suite for testing the Echofoam Theory under falsifiability conditions."
+authors = [{ name = "James Lorello", email = "your-email@example.com" }]
+readme = "README.md"
+requires-python = ">=3.8"
+license = { text = "MIT" }
+
+[project.urls]
+Homepage = "https://github.com/JamesLorello/Echofoam-Falsifiability-suite"
+Documentation = "https://github.com/JamesLorello/Echofoam-Falsifiability-suite/tree/main/docs"
+Source = "https://github.com/JamesLorello/Echofoam-Falsifiability-suite"
+Tracker = "https://github.com/JamesLorello/Echofoam-Falsifiability-suite/issues"
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
